More UI updates, reworking Resources/Costs/Rewards

+++ Assets/Scripts/UI/UIDiceSelection.cs
@@ -21,14 +21,14 @@ public class UIDiceSelection : MonoBehaviour {
 	
 	
 	public void SetupPanel() {
-		effortPointsRequired.text = quest.Reqs[(int)skill].RequiredValue.ToString();
-		effortPointsText.text = "of my <b>" + quest.Reqs[(int)skill].RequiredValue.ToString() + " " + skill.ToString() + "</b> dice.";
+		effortPointsRequired.text = quest.Reqs[skill].RequiredValue.ToString();
+		effortPointsText.text = "of my <b>" + quest.Reqs[skill].RequiredValue.ToString() + " " + skill.ToString() + "</b> dice.";
 		// TODO: Change the image skill here
 		effortType.sprite = icons[(int)skill];
 		for (int i = 1; i < player.Skills[skill].DiceTotal; i++) {
 			diceNumber.options.Add(new Dropdown.OptionData() { text = i.ToString() });	
 		}
-		string effortMessage = "This task requires <b>" + quest.Reqs[(int)skill].RequiredValue.ToString() + " " + skill.ToString() + " Effort Points</b> to complete.";
+		string effortMessage = "This task requires <b>" + quest.Reqs[skill].RequiredValue.ToString() + " " + skill.ToString() + " Effort Points</b> to complete.";
 		effortPointsText.text = effortMessage;
 	}
